+++ src/lib.rs
@@ -198,6 +198,7 @@ pub mod core_animation;
 pub mod core_foundation;
 pub mod core_graphics;
 pub mod core_image;
+pub mod core_video;
 pub mod dispatch;
 pub mod foundation;
 pub mod ui_kit;